requirements.txt
alembic==1.8.0
Babel==2.10.3
blinker==1.4
certifi==2022.6.15
charset-normalizer==2.0.12
click==8.1.3
colorama==0.4.4
dnspython==2.2.1
dominate==2.6.0
elastic-transport==8.1.2
elasticsearch==8.2.3
email-validator==1.2.1
Flask==2.1.2
Flask-Babel==2.0.0
Flask-Login==0.6.1
Flask-Mail==0.9.1
Flask-Migrate==3.1.0
Flask-Moment==1.0.2
Flask-SQLAlchemy==2.5.1
Flask-WTF==1.0.1
greenlet==1.1.2
idna==3.3
importlib-metadata==4.11.4
importlib-resources==5.7.1
itsdangerous==2.1.2
Jinja2==3.0.3
langdetect==1.0.9
Mako==1.2.0
MarkupSafe==2.1.1
PyJWT==2.4.0
python-dotenv==0.20.0
pytz==2022.1
requests==2.28.0
six==1.16.0
SQLAlchemy==1.4.37
urllib3==1.26.9
visitor==0.1.3
Werkzeug==2.1.2
WTForms==3.0.1
zipp==3.8.0
